hierarchical focus legwork, gimmicky window transparency hack
Some steps towards properly integrating the vscode extension's input,
interrupted in the middle to support using the slider to mess with
opacity on windows using sliders.

Right now the 4 sliders control the opacity of the top 4 windows in
the current bank.  The expected use of the sliders is probably volume
control management.