vscode client works, with controller handling WM/non-wm clients and their interaction
In order for the vscode client sending us information to be useful, the
controller-server had to have an idea of how non-WM clients relate to
the windows, and to know to use their information when their windows
were focused.

Much of the hassle here was just that we need to tunnel the process.pid
from the vscode extension through the editor window in order to establish
the mapping.  (And that this was not the obvious first solution, with the
better-seeming solutions involving research that did not pan out.)

The net result is that the vscode editor reports its "columns" as focus
slots, reporting the documents as containers whose id's are relative paths
to the workspace root.  We now also store the focus slots on bookmarks as
well.  This means bookmarks for the left pane and right pane in vscode
are logically distinct.  This seems right for a first step, with it being
possible for controllers to introduce a "just use my current focus slot"
mode as a boolean mode-toggle or just specialized press (like long-press
or pressure-level press for a pressure-sensitive controller, etc.).