Update button color determination and assignment.
Color assignment pre-dated the "window" and "thing" granularities of
bookmark assignment.  This commit updates it so that now "quant" (which
corresponds to window granularity) is paired with "reverse" for color
assignment.  And "capture" is paired with "type" for color assignment.

I also updated the logic that determines bookmark visibility to return
"focused" for window-level bookmarks.  Previously these would never be
marked as focused.