hack around technical debt of insufficiently cleaned up window meta ids
An undefined de-ref happened after restarting Firefox nightly.  I think
I'd originally punted on cleaning up the meta because I was just trying
to make things work.  Not entirely sure if I implemented
rebuild-on-disconnect.  If I did, it clearly was insufficient or has an
edge case.  Either way it needs to be looked into and fixed.