color computation refactored to get f1 working
This cleans up the split between RGB and indexed colors somewhat.
This leaves the F1 seemingly operational, mk3 is up next, but it
didn't work yet to begin with, so it will get cleaned up and
operational at the same time.

NB: ColorHelper is currently exported as a singleton and it's structurally
a bit weird.  I'd like to change it so the server just creates it and passes
it down.