minor null deref issue after trying to be case-insensitive.
I will also admit to not seeing whether role's contents could even be
anything other than lowercase.

I bumped the version by 0.0.2 instead of 0.0.1 because the AMO buttons
finally outwitted me.  (I always want to click on the abort button because
it's on the right and red... the correct button is blue and on the left.)