provide mechanism for clickable log entries to get log details
The 'client last message' dump box can now display different things
depending on a listbar above it.  A log details dump is the new
option and it allows log messages to have the simple single line
display that already existed, plus now you can provide a JS object
that will be pretty-JSON-serialized and displayed in the dump window.

The first user of this is the window mapping logic which originally
had gained very spammy console.log output that, in conjunction with
the other spammy outputs as clients told us things, made it impossible
to usefully figure out what was going wrong.  Now we can click on the
log entry where something went wrong to better understand what happened.

Especially since the detail objects can entrain a lot of memory, the
log has also been capped at 300 lines of scrollback and matching detail
info.  But it hasn't been tested, so it would not be surprising for it
to be broken somehow.