additional efforts to make atom clients to link focus slot at connection time
The general problem is that when atom clients send their focus slot inventories
we haven't managed to extract a PID from their window from the WM, apparently.
This means we fail to link the focus slots to their window container, which is
necessary for the new style of window bookmarking to work.

The efforts here were to:
- Further expand on buffering so that non-WM clients wouldn't be released to
  process until the initial WM client thingsExist report arrived and was
  processed, ideally with the PID slurped out of the titles.
- Makes the atom client attempt to be more proactive about tunneling the PID.
  This isn't perfect because the way we do it in atom is non-persistent (vs.
  in vscode) so it gets clobbered on tab changes.

Next step is probably to add some more event-based updates for:
- the focus slot being resolved to a window.  (I think this impacts tabs on top
  too.)
- trackWindow learning about the PID.  I'd originally punted on a full dep
  graph, but that might be what we're headed towards.