shrink and change font slightly
16pt was not giving enough horizontal space and vertically there was no
chance of 3 lines per grid cell.  I think we can do 3 lines now.

I've switched to sans-serif at the same time because serif was suffering
from the loss of pixels somewhat severely.  The sans-serif is better, but
not perfect.  I think the reduction will necessitate a change to the
grid background colors to further increase contrast.