bound the number of buffered conn messages to avoid OOM
The server has started OOMing on paintDisplay invocations, but I think this
is likely to be the straw that breaks the camel's back as opposed to the
actual problem.  (Flow control is in place for HTML rendering, and the message
debug mechanism should not be entraining rendered HTML messages.)

We'll now disconnect clients that have too many messages buffered, which is
a possible situation due to the window manager client disconnecting on screen
lock.  It's then possible for Firefox clients to spam the server due to sites
like discord which insist on triggering updates non-stop.  (There might need
to be some type of title idempotency added to the firefox client... I'm not
sure if discord is just constantly tapping document.title or what.)