various socket leak cleanup attempts
Looking at recent libstream commits, I think it may be the case that the
websocket connect method simply leaks if there's a connection failure.
This may be fixed in the most recent libsoup release or in the current
trunk, wherein cleanups have been made at least related to post-connect
negotiation.  It might also be an edge-case relatd to the gjs bindings.
Or maybe I'm just somehow leaking closures.

The most practical next steps are likely some combination of:
1. Issue a straightforward http request against the local server and only
   attempt to establish a websocket connection in that case.
2. Upgrade to a more modern gnome-shell PPA or even locally built.  If
   that doesn't fix it, it should at least be more straightforward to
   attach a debugger and/or rr and investigate and fix.