streamdeck initial changes; webextension made more generic
Some work towards allowing a streamdeck to be used as an alternate
interface.  The preliminary plan is to have it duplicate the task
slot mechanism.  The current use of the knob touch sensitivity to
show the current tasks on the mk3 feels like too much of a mode
change.

The streamdeck's button targets are too small to use the buttons on a
1-to-1 basis, but aggregating 4x1 or 2x2 or 3x2 or 4x2 could work as
buttons with a large display area.  (I'd mash with left hand rather
than shifting to an index finger single-touch or single thumb.  Middle
finger then likely becomes the activation, but pinky/ring/middle could
also end up mashing a 3x1 area.)

The manifest bump to 0.7.0 for the webext hasn't been published yet.